+++ Engine/src/TileLayer.cpp
@@ -57,11 +57,11 @@ void CTileLayer::Draw()
 
 STileset CTileLayer::GetTilesetByID(int tileID)
 {
-	for (int i = 0; i < m_tilesets.size(); i++)
+	for (size_t i = 0; i < m_tilesets.size(); i++)
 	{
-		if (i + 1 <= m_tilesets.size() - 1)
+		if (i + (size_t)1 <= m_tilesets.size() - 1)
 		{
-			if (tileID >= m_tilesets[i].FirstGID && tileID < m_tilesets[i + 1].FirstGID)
+			if (tileID >= m_tilesets[i].FirstGID && tileID < m_tilesets[i + (size_t)1].FirstGID)
 			{
 				return m_tilesets[i];
 			}
